Flipflop avx2

See merge request algorithm/taiyaki!277

+++ taiyaki/ctc/c_crf_flipflop.h
@@ -1,11 +1,11 @@
 #include <stdint.h>
 
 void crf_flipflop_grad(float const *logprob, size_t ntrans, size_t nblk,
-                       size_t nbatch, int32_t const *moveidxs,
-                       int32_t const *stayidxs, int32_t const *seqlen,
+                       size_t nbatch, size_t const *moveidxs,
+                       size_t const *stayidxs, int32_t const *seqlen,
                        float sharpfact, float *score, float *grad);
 
 void crf_flipflop_cost(float const *logprob, size_t ntrans, size_t nblk,
-                       size_t nbatch, int32_t const *moveidxs,
-                       int32_t const *stayidxs, int32_t const *seqlen,
+                       size_t nbatch, size_t const *moveidxs,
+                       size_t const *stayidxs, int32_t const *seqlen,
                        float sharpfact, float *score);
